+++ packages/node-xlsx/src/xlsxStream.ts
@@ -1,0 +1,10 @@
+class XlsxStream {
+
+    public async getHeaderTypes(filePath: string): Promise<string> {
+
+
+        return "";
+    }
+}
+
+export const xlsxStream = new XlsxStream();