+++ packages/fluent-firebase/CHANGELOG.md
@@ -1,5 +1,14 @@
 # 0.5.20
 
+## 0.7.18
+
+### Patch Changes
+
+- 6ec8e94: Update to pnpm
+- Updated dependencies [6ec8e94]
+  - @goatlab/js-utils@0.8.34
+  - @goatlab/fluent@0.7.32
+
 ## 0.7.17
 
 ### Patch Changes